style(core): separated http and non http error types
