+++ packages/core/src/index.ts
@@ -1,6 +1,7 @@
 // http
 export { EffectFactory } from './http/effects/http.effects.factory';
 export { defaultError$ } from './http/error/http.error.effect';
+export { HttpError } from './http/error/http.error.model';
 export { combineRoutes } from './http/router/http.router.combiner';
 export { r } from './http/router/http.router.ixbuilder';
 export * from './http/router/http.router.interface';
@@ -12,23 +13,23 @@ export * from './http/server/http.server.factory';
 export * from './http/server/http.server.listener';
 export * from './http/http.interface';
 
-// error
+// core - error
 export { coreErrorFactory, CoreErrorOptions } from './error/error.factory';
-export { HttpError, CoreError, EventError } from './error/error.model';
+export { CoreError, EventError } from './error/error.model';
 
-// effects
+// core - effects
 export { combineEffects, combineMiddlewares } from './effects/effects.combiner';
 export { createEffectContext } from './effects/effectsContext.factory';
 export * from './effects/effects.interface';
 
-// operators
+// core - operators
 export * from './operators';
 
-// event
+// core - event
 export * from './event/event.factory';
 export * from './event/event.interface';
 
-// context
+// core - context
 export * from './context/context.factory';
 export * from './context/context.hook';
 export * from './context/context.token.factory';
